+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.x branch
 ## 2.4 branch
 ### 2.4.5
-* now you can download logs.
+* now you can download and deletes logs.
 
 ### 2.4.4
 * fixed a major bug for the menu builder;