+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.4.5
 * logged users can view future posts, future pages and drafts;
 * the status of a record is properly shown in the admin panel, including publication in the future;
+* added filter for tags;
 * now you can restore database backups;
 * now you can download and deletes logs.